+++ tests/test_utils/test_core_utils.py
@@ -82,6 +82,15 @@ class UtilsTest(unittest.TestCase):
         self.assertFalse(pieces.areas)
 
 
+    def test_basepath(self):
+        """ """
+        test_path = '/some/path/to/test'
+
+        result = TmmsUtils.basepath(test_path, '/some/path/')
+        self.assertEqual(result, 'to/test')
+
+        self.assertRaises(AssertionError, TmmsUtils.basepath(test_path, '/path/to'))
+
     def touch_file(self, filename):
         """ Touch a file that will be used for testing. """
         with open(filename, 'a'):
