+++ utils/customize_node.py
@@ -31,8 +31,9 @@ from tmms.utils.logging import tmmsLogger
 from tmms.utils.file_utils import copy_target_into, remove_target
 from tmms.utils.file_utils import make_symlink, move_target
 from tmms.utils.file_utils import write_to_file, workdir, make_dir
-from tmms.utils.utils import find, piper, untar, kill_chroot_daemons
+#from tmms.utils.utils import find, piper, untar, kill_chroot_daemons
 from tmms.utils import utils # FIXME: this line should replace above imports!
+from tmms.utils import core_utils
 
 #==============================================================================
 # A (custom) kernel will probably only boot itself and not load any modules.
@@ -535,7 +536,7 @@ def create_cpio(args):
     update_status(args, 'Create %s from %s' % (cpio_file, args.new_fs_dir))
     try:
         # Skip things even though they may have been moved
-        found_data = find(
+        found_data = core_utils.find(
             args.new_fs_dir,
             ignore_files=['vmlinuz', 'initrd.img'],
             ignore_dirs=['boot'])
@@ -550,7 +551,7 @@ def create_cpio(args):
             # ./boot...). This causes Kernel Panic when trying to boot with
             # such a cpio file.
             with workdir(args.new_fs_dir):
-                ret, cpio_out, cpio_err = piper(
+                ret, cpio_out, cpio_err = core_utils.piper(
                     cmd, stdin=cpio_stdin, stdout=dest_obj)
                 assert not ret, 'cpio failed: %s' % cpio_err
 
@@ -751,14 +752,14 @@ echo 'LANG="en_US.UTF-8"' >> /etc/default/locale
     try:
         procmount = args.new_fs_dir + '/proc'
         os.makedirs(procmount, exist_ok=True)
-        ret, stdout, sterr = piper('mount -obind /proc ' + procmount)
+        ret, stdout, sterr = core_utils.piper('mount -obind /proc ' + procmount)
         assert not ret, 'Cannot bind mount /proc'
 
         umount = 'umount -fl %s' % procmount
 
         ptsmount = args.new_fs_dir + '/dev/pts'
         os.makedirs(ptsmount, exist_ok=True)
-        ret, stdout, sterr = piper('mount -obind /dev/pts ' + ptsmount)
+        ret, stdout, sterr = core_utils.piper('mount -obind /dev/pts ' + ptsmount)
         assert not ret, 'Cannot bind mount /dev/pts'
 
         umount = 'umount -fl %s %s' % (procmount, ptsmount)
@@ -773,7 +774,7 @@ echo 'LANG="en_US.UTF-8"' >> /etc/default/locale
         # This can take MINUTES, ie, "album" pulls in about 80 more packages.
         # While running, sys-images/nodeXX/untar/root/install.log is updated.
         # Hopefully install.sh catches its own errors
-        ret, stdout, stderr = piper(cmd, use_call=True)
+        ret, stdout, stderr = core_utils.piper(cmd, use_call=True)
         if ret:
             stdouterr = ''
             if stdout is not None:
@@ -789,8 +790,8 @@ echo 'LANG="en_US.UTF-8"' >> /etc/default/locale
             args.logger.debug(' - D - %s' % log.read())
         raise RuntimeError(str(err))
     finally:
-        umountret, _, _ = piper(umount)
-        kill_chroot_daemons(args.build_dir)
+        umountret, _, _ = core_utils.piper(umount)
+        utils.kill_chroot_daemons(args.build_dir)
     return False
 
 
@@ -893,14 +894,14 @@ def create_ESP(args, blockdev, vmlinuz, cpio):
             'Cannot find mapper file %s' % blockdev
 
         cmd = 'mkfs.vfat ' + blockdev
-        ret, stdout, stderr = piper(cmd)
+        ret, stdout, stderr = core_utils.piper(cmd)
         assert not ret, cmd
 
         # Step 3: Make, mount, and fill the VFAT FS.  The EFI default startup
         # script goes at /, but the grub stuff lives under "prefix".
 
         cmd = 'mount %s %s' % (blockdev, ESP_mnt)
-        ret, stdout, stderr = piper(cmd)
+        ret, stdout, stderr = core_utils.piper(cmd)
         assert not ret, cmd
         undo_mount = True
         update_status(args, 'SDHC GRUB DIR established at %s' % grubdir)
@@ -926,7 +927,7 @@ def create_ESP(args, blockdev, vmlinuz, cpio):
 
     if undo_mount:
         cmd = 'umount ' + ESP_mnt
-        ret, stdout, stderr = piper(cmd)
+        ret, stdout, stderr = core_utils.piper(cmd)
         assert not ret, cmd
 
     return undo_mount   # suppress final copy if this didn't work
@@ -967,14 +968,14 @@ def create_SNBU_image(args, vmlinuz, cpio):
         cmd += 'unit MiB mkpart primary fat32 %d 100%% ' % ESP_offset
         cmd += 'set 1 boot on set 1 esp on '
         cmd += 'name 1 %s ' % args.hostname
-        ret, stdout, stderr = piper(cmd)
+        ret, stdout, stderr = core_utils.piper(cmd)
         assert not ret, cmd
 
         # Step 2: located the partition and create a block device.
         # Ass-u-me enough loopback devics to go around.
 
         cmd = 'kpartx -asv %s' % ESP_img
-        ret, stdout, stderr = piper(cmd)
+        ret, stdout, stderr = core_utils.piper(cmd)
         assert not ret, cmd
         undo_kpartx = True
         stdout = stdout.decode()
@@ -1003,7 +1004,7 @@ def create_SNBU_image(args, vmlinuz, cpio):
         # server and "setnodes all".   Early in diagnosis, not sure what
         # to do about it.   Just move along for now and see what happens
         cmd = 'kpartx -d %s' % ESP_img
-        ret, stdout, stderr = piper(cmd)
+        ret, stdout, stderr = core_utils.piper(cmd)
         # assert not ret, cmd
         if ret or stderr:
             args.logger.warning('kpartx -d returned %d:\n - %s' % (ret, stderr))
@@ -1147,7 +1148,7 @@ def execute(args):
     # When some of them fail they'll handle last update_status themselves.
     try:
         update_status(args, 'Untar golden image')
-        args.new_fs_dir = untar(args.build_dir + '/untar/', args.golden_tar)
+        args.new_fs_dir = core_utils.untar(args.build_dir + '/untar/', args.golden_tar)
 
         set_foreign_package(args, 'qemu-aarch64-static')
